requirements.txt
Adafruit-DHT==1.4.0
blinker==1.9.0
click==8.2.1
colorzero==2.0
Flask==3.1.1
gpiozero==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
RPi.GPIO==0.7.1
Werkzeug==3.1.3
